% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark R and Rcpp functions.}
\description{
Use R package \code{microbenchmark} to compare the performance of C functions and Cpp functions
}
\examples{
\dontrun{
dt <- sample(1:100)
ts <- microbenchmark(bubble_sort=bubbleSort(dt),insert_sort = insertSort(dt),
select_sort = selectSort(dt),quick_sort=quickSort(dt),
bubble_sort_cpp=bubbleSortCpp(dt))
summary(ts)[,c(1,3,5,6)]
}
}
