% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bubbleSortCpp}
\alias{bubbleSortCpp}
\title{bubble sort algorithm using Rcpp}
\usage{
bubbleSortCpp(arr)
}
\arguments{
\item{arr}{Array to be sorted}
}
\value{
the sorted array
}
\description{
bubble sort algorithm using Rcpp
}
\examples{
\dontrun{
dt <- c(1,4,6,3,2);
bubbleSortCpp(dt);
}
}
