% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{myNeuralNet}
\alias{myNeuralNet}
\title{neural network}
\usage{
myNeuralNet(layers, activationFun)
}
\arguments{
\item{layers}{an array ,the i-th element in the array represents the number of nodes in the i-th layer of the neural network}

\item{activationFun}{choose an activation function}
}
\value{
Initial weight of network
}
\description{
construct fully connected neural network
}
\examples{
\dontrun{
myNeuralNet(c(13,6,3,2,1),'tanh')
}
}
