% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{prediction}
\alias{prediction}
\title{predict}
\usage{
prediction(x, fitWeights, activationFun)
}
\arguments{
\item{x}{sample}

\item{fitWeights}{Trained weight}

\item{activationFun}{activationFun}
}
\value{
prediction
}
\description{
input a sample , output predicted value
}
