% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{quickSort}
\alias{quickSort}
\title{quick sort algorithm}
\usage{
quickSort(arr, reverse = FALSE)
}
\arguments{
\item{arr}{the array to be sorted}

\item{reverse}{whether to sort in ascending or descending order}
}
\value{
an ordered array
}
\description{
quick sort algorithm
}
\examples{
\dontrun{
arr <- c(2,3,5,4,1)
quickSort(arr)
}
}
