% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{reluDerivate}
\alias{reluDerivate}
\title{relu function Derivative}
\usage{
reluDerivate(x)
}
\arguments{
\item{x}{independent variable}
}
\value{
dependent variable
}
\description{
relu function Derivative
}
\examples{
\dontrun{
reluDerivate(2)
}
}
