% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{reluFun}
\alias{reluFun}
\title{Relu function}
\usage{
reluFun(x)
}
\arguments{
\item{x}{independent variable}
}
\value{
dependent variable
}
\description{
Relu function
}
\examples{
\dontrun{
reluFun(2)
}
}
