% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{selectSort}
\alias{selectSort}
\title{select sort algorithm}
\usage{
selectSort(arr, reverse = FALSE)
}
\arguments{
\item{arr}{the array to be sorted}

\item{reverse}{whether to sort in ascending or descending order}
}
\value{
an ordered array
}
\description{
select sort algorithm
}
\examples{
\dontrun{
arr <- c(2,3,5,4,1)
selectSort(arr)
}
}
