% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{sigmoidDerivate}
\alias{sigmoidDerivate}
\title{sigmoid function Derivative}
\usage{
sigmoidDerivate(x)
}
\arguments{
\item{x}{independent variable}
}
\value{
dependent variable
}
\description{
sigmoid function Derivative
}
\examples{
\dontrun{
sigmoidDerivate(2)
}
}
