% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{sigmoidFun}
\alias{sigmoidFun}
\title{sigmoid function}
\usage{
sigmoidFun(x)
}
\arguments{
\item{x}{independent variable}
}
\value{
dependent variable
}
\description{
sigmoid function
}
\examples{
\dontrun{
sigmoidFun(2)
}
}
