\name{StatComp22092-package}
\alias{StatComp22092-package}
\alias{StatComp22092}
\docType{package}
\title{
  Compare the performance of four sorting algorithms and can construct fully connected neural network
}
\description{
  a R package developed to compare the performance of four sorting algorithms:bubble sort algorithm,selection sort algorithm,  insert sort algorithm ,quick sort algorithm; provides an interface to construct a neural network.
}
\author{
Wei Feng
Maintainer: Wei Feng <343024049@qq.com>
}
\keyword{ package }
