% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{swap}
\alias{swap}
\title{swap two numbers}
\usage{
swap(arr, i, j)
}
\arguments{
\item{arr}{the array to be swaped}

\item{i}{the first number index}

\item{j}{the second number index}
}
\value{
an  array
}
\description{
exchange two numbers
}
\examples{
\dontrun{
arr <- c(2,3,5,4,1)
swap(arr,1,2)
}
}
