% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{tanhDerivate}
\alias{tanhDerivate}
\title{tanh function Derivative}
\usage{
tanhDerivate(x)
}
\arguments{
\item{x}{independent variable}
}
\value{
dependent variable
}
\description{
tanh function Derivative
}
\examples{
\dontrun{
tanhDerivate(2)
}
}
