% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{training}
\alias{training}
\title{training the  network}
\usage{
training(x, y, learningRate, initialWeights, activationFun, epochs = 100)
}
\arguments{
\item{x}{training set}

\item{y}{labels}

\item{learningRate}{learning Rate}

\item{initialWeights}{Initial weight of network}

\item{activationFun}{activation function}

\item{epochs}{how many times to run the data}
}
\value{
Trained weight
}
\description{
use the traing set training the defined network
}
\examples{
\dontrun{
library(MASS)
samplesSelect<-sample(c(1:length(Boston[,1])),length(Boston[,1])*0.7)
trainData<-Boston[samplesSelect,]
testData<-Boston[-samplesSelect,]
trainX<-trainData[-which(colnames(Boston)=="medv")]
trainY<-trainData[which(colnames(Boston)=="medv")]
initialWeights <- myNeuralNet(c(13,6,3,2,1),'tanh')
fitWeights<-training(trainX,trainY,0.001,initialWeights,'tanh',20000)
}
}
